refactor: moving reagants to their own file, moving devices to devices
